databricks-delta-live-pipeline/requirements.txt
# PySpark used in real Databricks environment
pyspark

# YAML config handling
pyyaml

# Pandas used for local simulation
pandas